Refactor json schemas
