+++ app/api/templates/templateSchema.js
@@ -1,11 +1,9 @@
 import Ajv from 'ajv';
 import ajvKeywords from 'ajv-keywords';
-import { templateTypes } from 'shared/templateTypes';
 import model from 'api/templates/templatesModel';
+import { objectIdSchema, propertySchema } from 'api/utils/jsonSchemas'
 
 const ajv = ajvKeywords(Ajv({ allErrors: true }), ['uniqueItemProperties']);
-
-const fieldTypes = Object.values(templateTypes);
 
 ajv.addKeyword('uniqueName', {
   async: true,
@@ -100,46 +98,6 @@ ajv.addKeyword('requireInheritPropertyForInheritingRelationship', {
     return true;
   }
 });
-
-const objectIdSchema = {
-  oneOf: [
-    { type: 'string' },
-    { type: 'object' }
-  ]
-};
-
-const propertySchema = {
-  type: 'object',
-  required: ['label', 'type'],
-  requireContentForSelectFields: true,
-  requireRelationTypeForRelationship: true,
-  requireInheritPropertyForInheritingRelationship: true,
-  properties: {
-    id: { type: 'string' },
-    label: { type: 'string', minLength: 1 },
-    name: { type: 'string', minLength: 1 },
-    isCommonProperty: { type: 'boolean' },
-    type: { type: 'string', enum: fieldTypes },
-    prioritySorting: { type: 'boolean' },
-    content: { type: 'string', minLength: 1 },
-    inherit: { type: 'boolean' },
-    inheritProperty: { type: 'string', minLength: 1 },
-    filter: { type: 'boolean' },
-    noLabel: { type: 'boolean' },
-    fullWidth: { type: 'boolean' },
-    defaultfilter: { type: 'boolean' },
-    required: { type: 'boolean' },
-    sortable: { type: 'boolean' },
-    showInCard: { type: 'boolean' },
-    style: { type: 'string' },
-    nestedProperties: {
-      type: 'array',
-      items: {
-        type: 'string'
-      }
-    }
-  }
-};
 
 const schema = {
   $schema: 'http://json-schema.org/schema#',
